Refactor CraftManager to CraftController and update imports
Replaced occurrences of CraftManager with CraftController to better align with its function as a controller. Updated package structure and imports across affected files for consistency and clarity. This change improves maintainability and adheres to standard naming conventions.

Fixed Crafting System
